MoEncoder: Don't use separate velocity conversion factor
Previously, we were using a separate velocity conversion factor to account
for the different time bases of the velocity measures of different
encoders. That is, we would use setVelocityFactor to force the encoder
to always return velocity in units of internalEncoderUnits per second.

However, this is not supported by all encoder types (such as
CTRE), and I suspect it might be bugged on Rev when using SmartMotion.

So instead of using setVelocityFactor to cause the encoder to report
velocity in internalEncoderUnits per second, instead we just assume
the velocity will be returned by the encoder in internalEncoderUnits per
velocityBaseUnit (aka. internalEncoderVelocityUnits) and we create an
appropriate Measure instance.