Remove ArmSubsystem::limitArmPositionRequest
This method was behaving oddly in that it would disable smartmotion when
the mechanism was against either soft stop. I don't see anything
obviously wrong with the implementation, but I don't want to deep-dive
into the bug since we're limited on time.

The motor controller already handles soft limits, so there isn't much
need for me to do it here. I'll just remove it.